requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
WTForms==3.0.1
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
# psycopg2==2.9.7  # Удалено для Railway - используем SQLite
openpyxl==3.1.2
qrcode==7.4.2
Pillow==10.0.0
